+++ src/classes/UniversalCache.php
@@ -7,6 +7,7 @@
  *
  */
 namespace classes;
+use Predis\Client;
     /**
      * A class for generate universal cache file as key-value in traditional way (static files)
      * 
@@ -38,6 +39,42 @@ namespace classes;
          * MD5 Hash for filename cache
          */
         private static $hash = true;
+
+        /**
+         * If set to true then traditional filebased cache will change to use memory RAM with redis server.
+         */
+        private static $useredis = REDIS_ENABLE;
+
+        /**
+         * Open Redis Connection.
+         */
+        private static function openRedis(){
+            try {
+                return new Client(self::paramRedis(),self::optionRedis());
+            } catch (Exception $e) {
+                header("Content-type: application/json; charset=utf-8");
+                $data = [
+                    'status' => 'error',
+                    'code' => $e->getCode(),
+                    'message' => trim($e->getMessage())
+                ];
+                die(json_encode($data));
+            }
+        }
+
+        /**
+         * Set Redis parameter (This parameter can be set from config.php).
+         */
+        private static function paramRedis(){
+            return json_decode(REDIS_PARAMETER);
+        }
+
+        /**
+         * Set Redis option (This option can be set from config.php).
+         */
+        private static function optionRedis(){
+            return json_decode(REDIS_OPTION);
+        }
 
         /**
 		 * Convert string to valid UTF8 chars (slower but support ANSII)
@@ -73,22 +110,29 @@ namespace classes;
         /**
          * Determine is current key already cached or not
          * 
-         * @param cachetime = Set expired time in second. Default value is 300 seconds (5 minutes)
+         * @param cachetime = Set expired time in second (only works for filebased cache). Default value is 300 seconds (5 minutes)
          * 
          * @return bool
          */
         public static function isCached($key,$cachetime=300) {
             if (self::$runcache){
                 $file = self::filePath($key);
-                // check the expired file cache.
-                $mtime = 0;
-                if (file_exists($file)) {
-                    $mtime = filemtime($file);
-                }
-                $filetimemod = $mtime + $cachetime;
-                // if the renewal date is smaller than now, return true; else false (no need for update)
-                if ($filetimemod < time()) {
-                    return false;
+                if (self::$useredis){
+                    $redis = self::openRedis();
+                    if (empty($redis->get($file))){
+                        return false;
+                    }
+                } else {
+                    // check the expired file cache.
+                    $mtime = 0;
+                    if (file_exists($file)) {
+                        $mtime = filemtime($file);
+                    }
+                    $filetimemod = $mtime + $cachetime;
+                    // if the renewal date is smaller than now, return true; else false (no need for update)
+                    if ($filetimemod < time()) {
+                        return false;
+                    }
                 }
             } else {
                 return false;
@@ -105,8 +149,13 @@ namespace classes;
          */
         public static function loadCache($key) {
             $file = self::filePath($key);
-            if (file_exists($file)) {
-                return file_get_contents($file);
+            if (self::$useredis){
+                $redis = self::openRedis();
+                return $redis->get($file);
+            } else {
+                if (file_exists($file)) {
+                    return file_get_contents($file);
+                }
             }
             return "";
         }
@@ -116,13 +165,21 @@ namespace classes;
          * 
          * @param key = The Key name
          * @param value = input value to cache
+         *  @param redis_agecache = Set expired time in second (only works for Redis). Default value is 300 seconds (5 minutes)
          * 
          */
-        public static function writeCache($key,$value="") {
+        public static function writeCache($key,$value="",$redis_agecache=300) {
             if (!empty($key)) {
                 $file = self::filePath($key);
                 $content = '{"key":"'.$key.'","value":'.json_encode(self::safeConvertToUTF8($value)).',"timestamp":"'.date('Y-m-d h:i:s', time()).'"}';
-                if (self::$runcache) file_put_contents($file, $content, LOCK_EX);
+                if (self::$runcache) {
+                    if (self::$useredis){
+                        $redis = self::openRedis();
+                        $redis->setex($file,$redis_agecache,$content);
+                    } else {
+                        file_put_contents($file, $content, LOCK_EX);
+                    }
+                }
             }
         }
 
@@ -136,13 +193,18 @@ namespace classes;
         public static function deleteCache($key,$agecache=0) {
             if (!empty($key)) {
                 $file = self::filePath($key);
-                if (file_exists($file)){
-                    if ($agecache=0){
-                        unlink($file);
-                    } else {
-                        $now   = time();
-                        if ($now - filemtime($file) >= $agecache) {
+                if (self::$useredis){
+                    $redis = self::openRedis();
+                    $redis->del($file);
+                } else {
+                    if (file_exists($file)){
+                        if ($agecache=0){
                             unlink($file);
+                        } else {
+                            $now   = time();
+                            if ($now - filemtime($file) >= $agecache) {
+                                unlink($file);
+                            }
                         }
                     }
                 }
@@ -289,8 +351,22 @@ namespace classes;
                 'bytes'=>[
                     'cache'=>['use'=>$size,'free'=>$free],
                     'hdd'=>['use'=>$usehdd,'free'=>$free,'total'=>$total]
-                ]
+                ],
+                'redis' =>self::getRedisInfo()
             ];
+        }
+
+        /**
+         * Get info Redis Server
+         * 
+         * @return array
+         */
+        public static function getRedisInfo() {
+            $redis = self::openRedis();
+            foreach ($redis as $node) {
+                $info = $node->info();
+            }
+            return $info;
         }
 
         /**
